# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ChartLogs Trading Journal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trades:
    get:
      operationId: listTrades
      tags: [trades]
      summary: List trades with filters
      parameters:
        - name: symbol
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [long, short]
        - name: outcome
          in: query
          schema:
            type: string
            enum: [win, loss, breakeven]
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: tag
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of trades
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeListResponse"

    post:
      operationId: createTrade
      tags: [trades]
      summary: Create a new trade
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradeInput"
      responses:
        "201":
          description: Trade created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /trades/{id}:
    get:
      operationId: getTrade
      tags: [trades]
      summary: Get a trade by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trade found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateTrade
      tags: [trades]
      summary: Update a trade
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradeUpdate"
      responses:
        "200":
          description: Trade updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteTrade
      tags: [trades]
      summary: Delete a trade
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # JOURNAL
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Trade:
      type: object
      required: [id, userId, symbol, type, entryPrice, positionSize, openTime, source, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        symbol:
          type: string
        type:
          type: string
          enum: [long, short]
        entryPrice:
          type: number
        exitPrice:
          type: ["number", "null"]
        positionSize:
          type: number
        stopLoss:
          type: ["number", "null"]
        takeProfit:
          type: ["number", "null"]
        openTime:
          type: string
        closeTime:
          type: ["string", "null"]
        pnl:
          type: ["number", "null"]
        pips:
          type: ["number", "null"]
        rrRatio:
          type: ["number", "null"]
        fees:
          type: ["number", "null"]
        source:
          type: string
          enum: [manual, mt4, mt5, csv]
        tags:
          type: array
          items:
            type: string
        emotion:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        screenshots:
          type: array
          items:
            type: string
        outcome:
          type: ["string", "null"]
          enum: [win, loss, breakeven, null]
        createdAt:
          type: string
    TradeInput:
      type: object
      required: [symbol, type, entryPrice, positionSize, openTime, source]
      properties:
        symbol:
          type: string
        type:
          type: string
          enum: [long, short]
        entryPrice:
          type: number
        exitPrice:
          type: ["number", "null"]
        positionSize:
          type: number
        stopLoss:
          type: ["number", "null"]
        takeProfit:
          type: ["number", "null"]
        openTime:
          type: string
        closeTime:
          type: ["string", "null"]
        fees:
          type: ["number", "null"]
        source:
          type: string
          enum: [manual, mt4, mt5, csv]
        tags:
          type: array
          items:
            type: string
        emotion:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        screenshots:
          type: array
          items:
            type: string
    TradeUpdate:
      type: object
      properties:
        symbol:
          type: string
        type:
          type: string
          enum: [long, short]
        entryPrice:
          type: number
        exitPrice:
          type: ["number", "null"]
        positionSize:
          type: number
        stopLoss:
          type: ["number", "null"]
        takeProfit:
          type: ["number", "null"]
        openTime:
          type: string
        closeTime:
          type: ["string", "null"]
        fees:
          type: ["number", "null"]
        source:
          type: string
          enum: [manual, mt4, mt5, csv]
        tags:
          type: array
          items:
            type: string
        emotion:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        screenshots:
          type: array
          items:
            type: string
    TradeListResponse:
      type: object
      required: [trades, total]
      properties:
        trades:
          type: array
          items:
            $ref: "#/components/schemas/Trade"
        total:
          type: integer

    # JOURNAL
...